Initial Frostball Implementation
Using the fireball as an example, I have added the ability to see a
frostball powerup icon and a frostball projectile.  I have added the
Frostball class in the projectile folder and  FrostballPowerup under the
Powerups folder.

I have also used 0x2 in the enum to indicate the frostball powerup,
which should be changed according to whatever number it should fall
under when all of the powerups come together.

For this code I have disabled the fireball projectile in order to fire
the frostball projectile.  I may need to add code later to replace
certain powerups when they become implemented.