Committing my  Jet Minion Enemy
He is fully implemented and even shows damage after he's been hit too
many times as well as retreating after a certain period of time on the
screen. At the moment I'm working on an animation and explosion class,
I've included them but they are still broken. Other than that the only
thing I have left to say is that I've made quite a few constant and
static variables that can be used to tweak the behaviour of the A.I.,
these can be found at the top of the class